refactor: :recycle: responsive폴더내의 mobile tablet desktop -> device로 합치기
