+++ app/src/pages/Home.tsx
@@ -1,7 +1,7 @@
 import { gql } from '@/__generated__';
 import { useQuery } from '@apollo/client';
 import styled from '@emotion/styled';
-import { AboveTablet, Mobile } from '@/utils/responsive';
+import { AboveTablet, Mobile } from '@/utils/responsive/Device';
 import {
   DesktopDashboard,
   MobileDashboard,
