+++ app/src/pages/Login.tsx
@@ -1,7 +1,7 @@
 import { AppLogoTitle } from '@/components/elements/AppLogoTitle';
 import { AppIntroText } from '@/components/elements/Login/AppIntroText';
 import { Divider, HStack, VStack } from '@/components/common';
-import { AboveTablet, Mobile } from '@/utils/responsive';
+import { AboveTablet, Mobile } from '@/utils/responsive/Device';
 import styled from '@emotion/styled';
 import { GoogleLogin } from '@react-oauth/google';
 import { Helmet } from 'react-helmet-async';
