--- app/src/utils/responsive/Desktop.tsx
@@ -1,10 +1,0 @@
-import { PropsWithChildren } from 'react';
-import { useMediaQuery } from 'react-responsive';
-
-export const useIsDesktop = () => useMediaQuery({ minWidth: 1280 });
-
-export const Desktop = ({ children }: PropsWithChildren) => {
-  const isDesktop = useIsDesktop();
-
-  return <>{isDesktop ? children : null}</>;
-};
