--- app/src/utils/responsive/index.ts
@@ -1,3 +1,0 @@
-export * from './Desktop';
-export * from './Tablet';
-export * from './Mobile';
